#3 #30 : remove verification of compatibility of configurations between the current config and the output folder
simplification : If this is a new configuration of SlicerCART (new or template), require empty output folder. Otherwise, we've taken the exact configuration and are working in the same directory as the selected output folder. In Issue #30, it can be revisited whether we would like to allow modifications to the configuration after having started the data collection (segmentation or classification) and which of these modifications should be allowed.